add MacOS build script
This is a bit messy, it might be better if I used Bash or Zsh instead of AppleScript, but it works. 
To use it, install python with brew, create a venv, install all needed modules for prakticalc using pip3, and use the osascript command to run the build script within the venv. Also make sure that python3 is the python installed via brew, not the system python.